进度条完全体

+++ NCLCore/Info.cs
@@ -5,7 +5,7 @@ namespace NCLCore
     public class Info
     {
         public string msg;
-        public int process;
+        public double process;
         public string TYPE = "info";
         private readonly ILog log = LogManager.GetLogger("Info");
         public Info(string msg, string TYPE)
@@ -14,7 +14,7 @@ namespace NCLCore
             this.TYPE = TYPE;
             log.Debug("[" + TYPE + "]" + msg);
         }
-        public Info(int process,string msg)
+        public Info(double process,string msg)
         {
             this.process = process;
             this.msg = msg;
